back/requirements.txt
# Django核心
Django>=4.2.0,<4.3.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
django-filter>=23.2
drf-yasg>=1.21.5
drf-nested-routers>=0.93.4
python-dotenv>=1.0.0

# 数据库
psycopg2-binary>=2.9.0 

# 图像处理
Pillow>=9.5.0

# 工具
pandas>=2.0.2
jieba>=0.42.1
scikit-learn>=1.2.2
numpy>=1.24.3

# 部署
gunicorn>=21.2.0